requirements.txt
absl-py==0.14.1
alembic==1.4.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
cloudpickle==2.0.0
cycler==0.10.0
databricks-cli==0.15.0
docker==5.0.3
entrypoints==0.3
Flask==1.1.4
Flask-SocketIO==4.0.0
flatbuffers==1.12
gast==0.3.3
gensim==4.1.2
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mlflow==1.14.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.2.5
Pillow==8.3.2
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-editor==1.0.4
python-engineio==3.2.0
python-socketio==4.0.0
pytz==2021.3
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
smart-open==5.2.1
smmap==4.0.0
socketIO-client==0.7.2
SQLAlchemy==1.4.25
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.12.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.0.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.59.0
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==1.0.1
wrapt==1.12.1
xlwt==1.3.0
